requirements.txt
Flask==2.0.2
google-cloud-datastore==2.4.0
pandas==1.4.0
pytz==2021.3
matplotlib==3.5.1
google-cloud-storage==2.1.0